requirements.txt
alembic==1.1.0
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.21.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
passlib==1.7.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
redis==3.3.8
rq==1.1.0
six==1.12.0
SQLAlchemy==1.3.8
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
